calling functions in the jouer.py file from funtions.py file, details qdded in README.md
